Experimental tweaks
Try a new "incandescence" model that is actually hue in an attempt to
get a rainbow tree, but the RGB LEDs just don't seem to have enough
color resolution for it to look good...I can't quite figure it from the
library. There's a five-bit brightness value, but it's for every
channel, and for the library it's also common across every LED.

Tweak the bimetallic strip open temperature down a bit so LEDs
transition down into red. This is a bit of a stair-step, but increases
the visual interest a little bit too. This is the only effective change.

Also add tree.py to gitignore, for if copied or symlinked.